+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Added:
 - Ability to mark as read on buffer close only if scrolled to the bottom of the buffer (new default behavior)
 - `/part`, `/topic`, `/mode`, `/kick`, and `/ctcp` commands can have their target(s) argument skipped when it can be inferred from context (e.g. `/topic` will target the current channel by default when used in a channel buffer)
 - `/cleartopic` command to remove a channel's topic (will target the current channel by default when used in a channel buffer)
+- Ability to hide and theme `kick` server messages
+
+Fixed:
+- When kicked from a channel the kick message will be broadcast in the server buffer (which remains open) as well as in the channel history (which is closed on kick)
 
 Fixed: